requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
async-timeout==4.0.3
attrs==24.3.0
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.0
cycler==0.12.1
dataclasses-json==0.6.7
distro==1.9.0
exceptiongroup==1.2.2
Flask==3.1.0
Flask-Cors==4.0.2
Flask-SQLAlchemy==3.1.1
fonttools==4.55.5
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.7
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.41
langchain-experimental==0.0.65
langchain-openai==0.1.25
langchain-pinecone==0.1.3
langchain-text-splitters==0.2.4
langchainhub==0.1.21
langcodes==3.3.0
langgraph==0.1.14
langsmith==0.1.137
MarkupSafe==3.0.2
marshmallow==3.26.0
matplotlib==3.9.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.2.1
numpy==1.26.4
openai==1.60.0
orjson==3.10.15
packaging==24.2
pandas==2.2.2
pillow==10.3.0
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
propcache==0.2.1
pydantic==2.10.6
pydantic_core==2.27.2
PyMySQL==1.1.1
pyparsing==3.2.1
pypdf==3.16.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==5.1.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
scipy==1.13.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
tenacity==8.5.0
tiktoken==0.8.0
tqdm==4.67.1
types-requests==2.31.0.6
types-urllib3==1.26.25.14
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==1.26.20
Werkzeug==3.1.3
yarl==1.18.3
zipp==3.21.0
